Share the example runner's transpiler and client with the other websites
`transformExampleModule`, `exampleRunnerAsset` and `ExampleRunnerClient`
were copied verbatim into ag-grid, ag-charts and ag-studio by the port of
the "serve the example page's scripts" work (#14855). They are the parts
of the example runner with no product in them, so they move to
`ag-website-shared` and the copies here are deleted.

Unlike the other two sites, examples here are built per framework and are
given only the compiler options they need: `jsx` would rewrite a plain
`.ts` example's angle brackets, and the decorator options are Angular's
alone. That policy is the one thing in the file that is really ours, so
it stays, as `exampleCompilerOptions.ts`, and is handed to the shared
`transformExampleModule` as `compilerOptionNames`.

`transformExampleModule.test.ts` moves with the code. Its framework-blind
cases run in the shared project; the two that turn on the per-framework
options -- JSX for React, decorator metadata for Angular -- move to
`exampleCompilerOptions.test.ts`, along with a case pinning that a plain
TypeScript example emits no decorator metadata.

One assertion in `browserTranspiler.test.ts` compared `getCompilerOptions`
against `resolveCompilerOptions` of the same names, which is now the same
call twice over. It instead pins that every name the page carries resolves
to a real TypeScript enum member, which is what it was guarding against.

No behaviour changes.
