+++ documentation/ag-grid-docs/src/components/example-runner/components/PostInitMessageScript.astro
@@ -4,7 +4,7 @@ import {
     POST_INIT_MESSAGE_END,
     LIBRARY_INIT_SELECTOR,
 } from '@components/example-runner/constants';
-import { ExampleRunnerCall } from '@components/example-runner/framework-templates/lib/ExampleRunnerClient';
+import { ExampleRunnerCall } from '@ag-website-shared/components/example-runner/components/ExampleRunnerClient';
 interface Props {
     pageName: string;
     exampleName: string;
