+++ documentation/ag-grid-docs/src/components/example-runner/framework-templates/JavascriptTemplate.astro
@@ -52,7 +52,10 @@ import {
     getChartsEnterpriseScriptPath,
     getGridLocaleScriptPath,
 } from '@utils/gridLibraryPaths';
-import { ExampleRunnerCall, ExampleRunnerClient } from './lib/ExampleRunnerClient';
+import {
+    ExampleRunnerCall,
+    ExampleRunnerClient,
+} from '@ag-website-shared/components/example-runner/components/ExampleRunnerClient';
 import { SeedRandom } from './lib/SeedRandom';
 
 const siteUrl = Astro.site ? pathJoin(Astro.site, SITE_BASE_URL) : '/';
