+++ documentation/ag-grid-docs/src/pages/examples/[pageName]/[exampleName]/[internalFramework]/[fileName].ts
@@ -1,8 +1,12 @@
 import type { InternalFramework } from '@ag-grid-types';
+import {
+    getModuleSourceFileName,
+    transformExampleModule,
+} from '@ag-website-shared/components/example-runner/utils/transformExampleModule';
 import { getDocExampleFiles } from '@components/docs/utils/pageData';
 import { getGeneratedContents } from '@components/example-generator';
 import { getIsDev } from '@utils/env';
-import { getModuleSourceFileName, transformExampleModule } from '@utils/exampleModules/transformExampleModule';
+import { getCompilerOptionNames } from '@utils/exampleModules/exampleCompilerOptions';
 import { fileNameToMimeType } from '@utils/mimeType';
 import { getContentRootFileUrl } from '@utils/pages';
 import { getCollection } from 'astro:content';
@@ -52,7 +56,7 @@ export async function GET({ params }: { params: Params }) {
         const code = transformExampleModule({
             fileName: moduleSourceFileName,
             source: files[moduleSourceFileName],
-            internalFramework,
+            compilerOptionNames: getCompilerOptionNames(internalFramework),
         });
 
         return new Response(code, {
